requirements_supervised.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
archive=0.3=pypi_0
argparse=1.4.0=pypi_0
attrs=22.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.4=py_1
boto3=1.24.51=pyhd8ed1ab_0
botocore=1.27.52=pyhd8ed1ab_0
bottleneck=1.3.5=py310ha9d4c09_0
bravado=11.0.3=pyhd8ed1ab_0
bravado-core=5.17.0=pyh9f0ad1d_0
brotli=1.0.9=pypi_0
brotlicffi=1.0.9.2=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.6.15=ha878542_0
certifi=2022.6.15=py310hff52083_0
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py310hff52083_0
cryptography=37.0.1=py310h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.30=pypi_0
dataclasses=0.8=pyhc8e2a94_3
decorator=5.1.1=pypi_0
docutils=0.19=pypi_0
easydict=1.10=pypi_0
easyprocess=1.1=pypi_0
eden=2.1.4=pypi_0
entrypoint2=1.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.34.2=pypi_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py310hff52083_5
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
googledrivedownloader=0.4=pypi_0
h5py=3.7.0=py310he06866b_0
hdf5=1.10.6=h3ffc7dd_1
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=py310hff52083_0
inflate64=0.1.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
isodate=0.6.1=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pyhd8ed1ab_0
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
jsonref=0.2=py_0
jsonschema=3.2.0=pyhd8ed1ab_3
kivy=2.1.0=pypi_0
kivy-garden=0.1.5=pypi_0
kiwisolver=1.4.3=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
monotonic=1.5=py_0
msgpack-python=1.0.3=py310hd09550d_0
multivolumefile=0.2.3=pypi_0
ncurses=6.3=h5eee18b_3
neptune-client=0.16.5=pyhd8ed1ab_1
nettle=3.7.3=hbbd107a_1
networkx=2.5=pypi_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.22.3=py310hfa59a62_0
numpy-base=1.22.3=py310h9585f30_0
oauthlib=3.2.0=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=pypi_0
patool=1.12=pypi_0
pillow=9.0.1=py310h22f2fdc_0
pip=21.2.4=py310h06a4308_0
psutil=5.9.1=py310h5764c6d_0
py7zr=0.19.0=pypi_0
pybcj=0.6.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.15.0=pypi_0
pyemd=0.5.1=pypi_0
pygments=2.12.0=pypi_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyppmd=0.18.3=pypi_0
pyrsistent=0.18.1=py310h5764c6d_1
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.16=pypi_0
python_abi=3.10=2_cp310
pytorch=1.12.0=py3.10_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pyunpack=0.3=pypi_0
pyyaml=6.0=py310h5764c6d_4
pyzstd=0.15.2=pypi_0
rdflib=6.1.1=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.0=py310h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
s3transfer=0.6.0=pyhd8ed1ab_0
scikit-learn=1.1.1=pypi_0
scipy=1.8.1=pypi_0
seaborn=0.11.2=pypi_0
setuptools=61.2.0=py310h06a4308_0
simplejson=3.17.6=py310h5764c6d_1
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.38.5=hc218d9a_0
swagger-spec-validator=2.7.4=pyhd8ed1ab_0
texttable=1.6.4=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch-geometric=1.7.1=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.14=pypi_0
torchaudio=0.12.0=py310_cu113
torchvision=0.13.0=py310_cu113
tqdm=4.64.0=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310h06a4308_0
websocket-client=1.3.3=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7f98852_2
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
